src/requirements.txt
msgpack==1.0.2
python-snappy==0.5.4
requests==2.24.0
